pyproject.toml
[tool.poetry]
name = "cat-rabbit-classifier"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.68.1"
Pillow = "^8.3.1"
uvicorn = "^0.15.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
